requirements.txt
requests
numpy
scipy
pillow
PyDrive
